check temp confirmation expiration on validate method and optimised cron

+++ src/main/java/com/rmr/dinosaurs/domain/auth/service/impl/TempConfirmationServiceImpl.java
@@ -1,5 +1,6 @@
 package com.rmr.dinosaurs.domain.auth.service.impl;
 
+import static com.rmr.dinosaurs.domain.auth.exception.errorcode.TempConfirmationErrorCode.CONFIRMATION_CODE_EXPIRED;
 import static com.rmr.dinosaurs.domain.auth.exception.errorcode.TempConfirmationErrorCode.CONFIRMATION_CODE_NOT_FOUND;
 
 import com.rmr.dinosaurs.domain.auth.model.TempConfirmation;
@@ -7,11 +8,11 @@ import com.rmr.dinosaurs.domain.auth.model.User;
 import com.rmr.dinosaurs.domain.auth.service.TempConfirmationService;
 import com.rmr.dinosaurs.domain.core.exception.ServiceException;
 import com.rmr.dinosaurs.infrastucture.database.auth.TempConfirmationRepository;
-import java.time.Instant;
 import java.time.LocalDateTime;
 import java.time.temporal.ChronoUnit;
 import java.util.Optional;
 import java.util.UUID;
+import java.util.function.BiPredicate;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.scheduling.annotation.Scheduled;
@@ -25,6 +26,8 @@ public class TempConfirmationServiceImpl implements TempConfirmationService {
   private static final Long TTL = 60L;
 
   private final TempConfirmationRepository tempConfirmationRepository;
+  private final BiPredicate<LocalDateTime, LocalDateTime> isValidTempConfirmationPredicate =
+      (currentTime, issuedAt) -> currentTime.minus(TTL, ChronoUnit.MINUTES).isAfter(issuedAt);
 
 
   @Override
@@ -40,19 +43,18 @@ public class TempConfirmationServiceImpl implements TempConfirmationService {
           log.info("Temp confirmation was not found");
           throw new ServiceException(CONFIRMATION_CODE_NOT_FOUND);
         });
+    if (!isValidTempConfirmationPredicate.test(LocalDateTime.now(),
+        tempConfirmation.getIssuedAt())) {
+      throw new ServiceException(CONFIRMATION_CODE_EXPIRED);
+    }
     tempConfirmationRepository.delete(tempConfirmation);
     return Optional.of(tempConfirmation);
   }
 
   @Scheduled(cron = "@midnight")
   private void removeNonConfirmed() {
-    // TODO: could be optimised to use just a query
-    var overDueTempConfirmations = tempConfirmationRepository.findAll()
-        .stream()
-        .filter(tmpConf -> Instant.now().minus(TTL, ChronoUnit.MINUTES).isBefore(
-            Instant.from(tmpConf.getIssuedAt())))
-        .toList();
-    tempConfirmationRepository.deleteAll(overDueTempConfirmations);
+    tempConfirmationRepository
+        .deleteAllByIssuedAtBefore(LocalDateTime.now().minus(TTL, ChronoUnit.MINUTES));
   }
 
 }
